+++ CHANGELOG.md
@@ -21,6 +21,7 @@ And constructed with the following guidelines:
 * Fixed the help contact popup window.
 * Fixed the field duplication issue for all summary reports.
 * Fixed the search functionality.  Added new report types to the search.
+* Updated the popups for supervisor and city council districts
 
 ## Version 3.2.0 | 01/18/2018